:package: routes configuration was extracted to separate file
